fix: silence remaining first-party Clang warnings
Eliminate all non-thirdparty warnings from the Clang warning gate
build. Remaining diagnostics are confined to thirdparty_unman/
(antiword, chmlib, qimagescale) which is not modified.

Changes span drawbuf, font, render, XML, DOCX/EPUB/PDB/WOL format
handlers, crgui, fb2props and locale data: unused parameters commented
out, dead variables/functions removed, sign-compare casts added,
sprintf replaced with snprintf, overloaded-virtual hiding fixed with
using-declarations, strict-prototype declarations corrected, and
unused private fields removed.

+++ crengine/src/lvfont/lvfreetypefontman.cpp
@@ -130,11 +130,11 @@ LVHashTable<lString8, font_lang_compat>* getSupportedLangs(FT_Face face) {
                 if (inRange && codePoint < second) {
                     codePoint++;
                 } else {
-                    if (i >= rec->char_set_sz)
+                    if ((unsigned int)i >= rec->char_set_sz)
                         break;
                     tmp = rec->char_set[i];
                     if (FC_LANG_START_INTERVAL_CODE == tmp) {       // code of start interval
-                        if (i + 2 < rec->char_set_sz) {
+                        if ((unsigned int)i + 2 < rec->char_set_sz) {
                             i++;
                             first = rec->char_set[i];
                             i++;
